+++ CHANGELOG.md
@@ -3,8 +3,8 @@ Change log
 
 For detailed information check [Blogophon's releases on Github](https://github.com/fboes/blogophon/releases).
 
+* [x] Adding article type "Review"
 * [x] Improved drafting of articles
-* [x] Improved Promises in Generator
 * [x] Image sublines for images
 
 1.2.0